Remove flag for plp, only check for subscription plan
