+++ apps/studio/components/interfaces/Organization/TeamSettings/InviteMemberButton.tsx
@@ -15,7 +15,10 @@ import { useOrganizationRolesV2Query } from 'data/organization-members/organizat
 import { useOrganizationMemberInviteCreateMutation } from 'data/organizations/organization-member-invite-create-mutation'
 import { useOrganizationMembersQuery } from 'data/organizations/organization-members-query'
 import { useProjectsQuery } from 'data/projects/projects-query'
-import { useOrgSubscriptionQuery } from 'data/subscriptions/org-subscription-query'
+import {
+  useHasAccessToProjectLevelPermissions,
+  useOrgSubscriptionQuery,
+} from 'data/subscriptions/org-subscription-query'
 import {
   doPermissionsCheck,
   useCheckPermissions,
@@ -58,7 +61,6 @@ import {
 } from 'ui'
 import { FormItemLayout } from 'ui-patterns/form/FormItemLayout/FormItemLayout'
 import { useGetRolesManagementPermissions } from './TeamSettings.utils'
-import { useIsOptedIntoProjectLevelPermissions } from 'hooks/ui/useFlag'
 
 export const InviteMemberButton = () => {
   const { slug } = useParams()
@@ -87,8 +89,7 @@ export const InviteMemberButton = () => {
     { enabled: canReadSubscriptions }
   )
   const currentPlan = subscription?.plan
-  const isOptedIntoProjectLevelPermissions =
-    useIsOptedIntoProjectLevelPermissions(slug as string) && currentPlan?.id === 'enterprise'
+  const hasAccessToProjectLevelPermissions = useHasAccessToProjectLevelPermissions(slug as string)
 
   const userMemberData = members?.find((m) => m.gotrue_id === profile?.gotrue_id)
   const hasOrgRole =
@@ -149,7 +150,7 @@ export const InviteMemberButton = () => {
       }
     }
 
-    if (isOptedIntoProjectLevelPermissions) {
+    if (hasAccessToProjectLevelPermissions) {
       inviteMember(
         {
           slug,
@@ -247,7 +248,7 @@ export const InviteMemberButton = () => {
             onSubmit={form.handleSubmit(onInviteMember)}
           >
             <DialogSection className="flex flex-col gap-y-4 pb-2">
-              {isOptedIntoProjectLevelPermissions && (
+              {hasAccessToProjectLevelPermissions && (
                 <FormField_Shadcn_
                   name="applyToOrg"
                   control={form.control}
