+++ apps/studio/hooks/ui/useFlag.ts
@@ -1,8 +1,9 @@
 import { PermissionAction } from '@supabase/shared-types/out/constants'
+import { useContext } from 'react'
+
 import FlagContext from 'components/ui/Flag/FlagContext'
 import { useOrgSubscriptionQuery } from 'data/subscriptions/org-subscription-query'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
-import { useContext } from 'react'
 
 export function useFlag<T = boolean>(name: string) {
   const store: any = useContext(FlagContext)
@@ -16,26 +17,4 @@ export function useFlag<T = boolean>(name: string) {
     return false
   }
   return store[name] as T
-}
-
-export const useIsOptedIntoProjectLevelPermissions = (slug: string) => {
-  const orgsOptedIn = useFlag('projectLevelPermissionsOptIn') as string | undefined
-  const canReadSubscriptions = useCheckPermissions(
-    PermissionAction.BILLING_READ,
-    'stripe.subscriptions'
-  )
-  const { data: subscription } = useOrgSubscriptionQuery(
-    { orgSlug: slug },
-    { enabled: canReadSubscriptions }
-  )
-
-  const isEnterprise = subscription?.plan.id === 'enterprise'
-  if (!isEnterprise) return false
-
-  if (typeof orgsOptedIn === 'string') {
-    const organizations = orgsOptedIn.split(',').map((x) => x.trim())
-    return organizations.includes(slug)
-  } else {
-    return false
-  }
 }
